+++ ErsatzTV.Scanner/Core/Interfaces/Metadata/ILocalChaptersProvider.cs
@@ -1,0 +1,8 @@
+using ErsatzTV.Core.Domain;
+
+namespace ErsatzTV.Scanner.Core.Interfaces.Metadata;
+
+public interface ILocalChaptersProvider : IDisposable
+{
+    Task<bool> UpdateChapters(MediaItem mediaItem, Option<string> localPath);
+}