+++ ErsatzTV.Scanner/Core/Plex/PlexOtherVideoLibraryScanner.cs
@@ -4,6 +4,7 @@ using ErsatzTV.Core.Extensions;
 using ErsatzTV.Core.Interfaces.Metadata;
 using ErsatzTV.Core.Interfaces.Plex;
 using ErsatzTV.Core.Interfaces.Repositories;
+using ErsatzTV.Scanner.Core.Interfaces.Metadata;
 using ErsatzTV.Core.Metadata;
 using ErsatzTV.Core.Plex;
 using ErsatzTV.Scanner.Core.Metadata;
@@ -32,9 +33,11 @@ public class PlexOtherVideoLibraryScanner :
         IPlexOtherVideoRepository plexOtherVideoRepository,
         IPlexPathReplacementService plexPathReplacementService,
         ILocalFileSystem localFileSystem,
+        ILocalChaptersProvider localChaptersProvider,
         ILogger<PlexOtherVideoLibraryScanner> logger)
         : base(
             localFileSystem,
+            localChaptersProvider,
             metadataRepository,
             mediator,
             logger)